IMO we can now merge TestsListView2 into default TestsListView
